backend/requirements.txt
asgiref==3.6.0
attrs==22.2.0
Django==3.2
django-cors-headers==3.14.0
django-filter==22.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
drf-spectacular==0.25.1
gunicorn==20.1.0
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.4.0
psycopg2-binary==2.9.5
pyrsistent==0.19.3
python-dotenv
pytz==2023.2
PyYAML==6.0
sqlparse==0.4.3
uritemplate==4.1.1
